+++ pyglet/graphics/api/gl/xlib/glx.py
@@ -19,7 +19,7 @@ if not _link_function:
 # This content is generated by tools/gengl.py.
 # Wrapper for /usr/include/GL/glx.h
 
-import pyglet.libs.x11.xlib
+import pyglet.libs.linux.x11.xlib
 
 # H (/usr/include/GL/glx.h:26)
 GLX_VERSION_1_1 = 1 	# /usr/include/GL/glx.h:58
@@ -119,7 +119,7 @@ struct___GLXcontextRec._fields_ = [
 ]
 
 GLXContext = POINTER(struct___GLXcontextRec) 	# /usr/include/GL/glx.h:178
-XID = pyglet.libs.x11.xlib.XID
+XID = pyglet.libs.linux.x11.xlib.XID
 GLXPixmap = XID 	# /usr/include/GL/glx.h:179
 GLXDrawable = XID 	# /usr/include/GL/glx.h:180
 class struct___GLXFBConfigRec(Structure):
@@ -134,8 +134,8 @@ GLXFBConfigID = XID 	# /usr/include/GL/glx.h:183
 GLXContextID = XID 	# /usr/include/GL/glx.h:184
 GLXWindow = XID 	# /usr/include/GL/glx.h:185
 GLXPbuffer = XID 	# /usr/include/GL/glx.h:186
-XVisualInfo = pyglet.libs.x11.xlib.XVisualInfo
-Display = pyglet.libs.x11.xlib.Display
+XVisualInfo = pyglet.libs.linux.x11.xlib.XVisualInfo
+Display = pyglet.libs.linux.x11.xlib.Display
 # /usr/include/GL/glx.h:190
 glXChooseVisual = _link_function('glXChooseVisual', POINTER(XVisualInfo), [POINTER(Display), c_int, POINTER(c_int)], 'H')
 
@@ -154,7 +154,7 @@ glXCopyContext = _link_function('glXCopyContext', None, [POINTER(Display), GLXCo
 # /usr/include/GL/glx.h:204
 glXSwapBuffers = _link_function('glXSwapBuffers', None, [POINTER(Display), GLXDrawable], 'H')
 
-Pixmap = pyglet.libs.x11.xlib.Pixmap
+Pixmap = pyglet.libs.linux.x11.xlib.Pixmap
 # /usr/include/GL/glx.h:206
 glXCreateGLXPixmap = _link_function('glXCreateGLXPixmap', GLXPixmap, [POINTER(Display), POINTER(XVisualInfo), Pixmap], 'H')
 
@@ -185,7 +185,7 @@ glXWaitGL = _link_function('glXWaitGL', None, [], 'H')
 # /usr/include/GL/glx.h:226
 glXWaitX = _link_function('glXWaitX', None, [], 'H')
 
-Font = pyglet.libs.x11.xlib.Font
+Font = pyglet.libs.linux.x11.xlib.Font
 # /usr/include/GL/glx.h:228
 glXUseXFont = _link_function('glXUseXFont', None, [Font, c_int, c_int, c_int], 'H')
 
@@ -213,7 +213,7 @@ glXGetFBConfigs = _link_function('glXGetFBConfigs', POINTER(GLXFBConfig), [POINT
 # /usr/include/GL/glx.h:254
 glXGetVisualFromFBConfig = _link_function('glXGetVisualFromFBConfig', POINTER(XVisualInfo), [POINTER(Display), GLXFBConfig], 'H')
 
-Window = pyglet.libs.x11.xlib.Window
+Window = pyglet.libs.linux.x11.xlib.Window
 # /usr/include/GL/glx.h:257
 glXCreateWindow = _link_function('glXCreateWindow', GLXWindow, [POINTER(Display), GLXFBConfig, Window, POINTER(c_int)], 'H')
