+++ pyglet/graphics/api/gl/xlib/glxext_arb.py
@@ -24,7 +24,7 @@ if not hasattr(ctypes, 'c_int64'):
 # This content is generated by tools/gengl.py.
 # Wrapper for http://www.opengl.org/registry/api/glxext.h
 
-import pyglet.libs.x11.xlib
+import pyglet.libs.linux.x11.xlib
 import pyglet.graphics.api.gl.xlib.glx
 
 # H (/usr/include/GL/glx.h:26)
@@ -257,7 +257,7 @@ GLX_GPU_NUM_SPI_AMD = 8616 	# GL/glxext.h:436
 GLX_CONTEXT_ES2_PROFILE_BIT_EXT = 4 	# GL/glxext.h:440
 # ARB_get_proc_address (GL/glxext.h:446)
 # SGIX_video_source (GL/glxext.h:450)
-XID = pyglet.libs.x11.xlib.XID
+XID = pyglet.libs.linux.x11.xlib.XID
 GLXVideoSourceSGIX = XID 	# GL/glxext.h:451
 # SGIX_fbconfig (GL/glxext.h:454)
 GLXFBConfigIDSGIX = XID 	# GL/glxext.h:455
@@ -287,7 +287,7 @@ class struct_anon_106(Structure):
         'height',
         'count',
     ]
-Display = pyglet.libs.x11.xlib.Display
+Display = pyglet.libs.linux.x11.xlib.Display
 GLXDrawable = glx.GLXDrawable
 struct_anon_106._fields_ = [
     ('type', c_int),
@@ -397,14 +397,14 @@ glXGetFBConfigAttribSGIX = _link_function('glXGetFBConfigAttribSGIX', c_int, [PO
 glXChooseFBConfigSGIX = _link_function('glXChooseFBConfigSGIX', POINTER(GLXFBConfigSGIX), [POINTER(Display), c_int, POINTER(c_int), POINTER(c_int)], 'SGIX_fbconfig')
 
 GLXPixmap = glx.GLXPixmap
-Pixmap = pyglet.libs.x11.xlib.Pixmap
+Pixmap = pyglet.libs.linux.x11.xlib.Pixmap
 # GL/glxext.h:680
 glXCreateGLXPixmapWithConfigSGIX = _link_function('glXCreateGLXPixmapWithConfigSGIX', GLXPixmap, [POINTER(Display), GLXFBConfigSGIX, Pixmap], 'SGIX_fbconfig')
 
 # GL/glxext.h:681
 glXCreateContextWithConfigSGIX = _link_function('glXCreateContextWithConfigSGIX', GLXContext, [POINTER(Display), GLXFBConfigSGIX, c_int, GLXContext, c_int], 'SGIX_fbconfig')
 
-XVisualInfo = pyglet.libs.x11.xlib.XVisualInfo
+XVisualInfo = pyglet.libs.linux.x11.xlib.XVisualInfo
 # GL/glxext.h:682
 glXGetVisualFromFBConfigSGIX = _link_function('glXGetVisualFromFBConfigSGIX', POINTER(XVisualInfo), [POINTER(Display), GLXFBConfigSGIX], 'SGIX_fbconfig')
 
@@ -441,7 +441,7 @@ PFNGLXSELECTEVENTSGIXPROC = CFUNCTYPE(None, POINTER(Display), GLXDrawable, c_ulo
 PFNGLXGETSELECTEDEVENTSGIXPROC = CFUNCTYPE(None, POINTER(Display), GLXDrawable, POINTER(c_ulong)) 	# GL/glxext.h:706
 # SGI_cushion (GL/glxext.h:709)
 GLX_SGI_cushion = 1 	# GL/glxext.h:710
-Window = pyglet.libs.x11.xlib.Window
+Window = pyglet.libs.linux.x11.xlib.Window
 # GL/glxext.h:712
 glXCushionSGI = _link_function('glXCushionSGI', None, [POINTER(Display), Window, c_float], 'SGI_cushion')
 
@@ -501,7 +501,7 @@ glXCopySubBufferMESA = _link_function('glXCopySubBufferMESA', None, [POINTER(Dis
 PFNGLXCOPYSUBBUFFERMESAPROC = CFUNCTYPE(None, POINTER(Display), GLXDrawable, c_int, c_int, c_int, c_int) 	# GL/glxext.h:774
 # MESA_pixmap_colormap (GL/glxext.h:777)
 GLX_MESA_pixmap_colormap = 1 	# GL/glxext.h:778
-Colormap = pyglet.libs.x11.xlib.Colormap
+Colormap = pyglet.libs.linux.x11.xlib.Colormap
 # GL/glxext.h:780
 glXCreateGLXPixmapMESA = _link_function('glXCreateGLXPixmapMESA', GLXPixmap, [POINTER(Display), POINTER(XVisualInfo), Pixmap, Colormap], 'MESA_pixmap_colormap')
