+++ pyglet/graphics/api/gl/xlib/glxext_nv.py
@@ -18,7 +18,7 @@ from pyglet.graphics.api.gl.lib import c_void
 # This content is generated by tools/gengl.py.
 # Wrapper for http://developer.download.nvidia.com/opengl/includes/glxext.h
 
-import pyglet.libs.x11.xlib
+import pyglet.libs.linux.x11.xlib
 import pyglet.graphics.api.gl.xlib.glx
 
 # H (/usr/include/GL/glx.h:26)
@@ -148,7 +148,7 @@ GLX_HYPERPIPE_ID_SGIX = 32816 	# GL/glxext.h:240
 # MESA_agp_offset (GL/glxext.h:243)
 # ARB_get_proc_address (GL/glxext.h:249)
 # SGIX_video_source (GL/glxext.h:256)
-XID = pyglet.libs.x11.xlib.XID
+XID = pyglet.libs.linux.x11.xlib.XID
 GLXVideoSourceSGIX = XID 	# GL/glxext.h:257
 # SGIX_fbconfig (GL/glxext.h:260)
 GLXFBConfigIDSGIX = XID 	# GL/glxext.h:261
@@ -178,7 +178,7 @@ class struct_anon_106(Structure):
         'x',
         'y',
     ]
-Display = pyglet.libs.x11.xlib.Display
+Display = pyglet.libs.linux.x11.xlib.Display
 GLXDrawable = pyglet.graphics.api.gl.xlib.glx.GLXDrawable
 struct_anon_106._fields_ = [
     ('type', c_int),
@@ -318,14 +318,14 @@ glXGetFBConfigAttribSGIX = _link_function('glXGetFBConfigAttribSGIX', c_int, [PO
 glXChooseFBConfigSGIX = _link_function('glXChooseFBConfigSGIX', POINTER(GLXFBConfigSGIX), [POINTER(Display), c_int, POINTER(c_int), POINTER(c_int)], 'SGIX_fbconfig')
 
 GLXPixmap = pyglet.graphics.api.gl.x11.glx.GLXPixmap
-Pixmap = pyglet.libs.x11.xlib.Pixmap
+Pixmap = pyglet.libs.linux.x11.xlib.Pixmap
 # GL/glxext.h:460
 glXCreateGLXPixmapWithConfigSGIX = _link_function('glXCreateGLXPixmapWithConfigSGIX', GLXPixmap, [POINTER(Display), GLXFBConfigSGIX, Pixmap], 'SGIX_fbconfig')
 
 # GL/glxext.h:461
 glXCreateContextWithConfigSGIX = _link_function('glXCreateContextWithConfigSGIX', GLXContext, [POINTER(Display), GLXFBConfigSGIX, c_int, GLXContext, c_int], 'SGIX_fbconfig')
 
-XVisualInfo = pyglet.libs.x11.xlib.XVisualInfo
+XVisualInfo = pyglet.libs.linux.x11.xlib.XVisualInfo
 # GL/glxext.h:462
 glXGetVisualFromFBConfigSGIX = _link_function('glXGetVisualFromFBConfigSGIX', POINTER(XVisualInfo), [POINTER(Display), GLXFBConfigSGIX], 'SGIX_fbconfig')
 
@@ -362,7 +362,7 @@ PFNGLXSELECTEVENTSGIXPROC = CFUNCTYPE(None, POINTER(Display), GLXDrawable, c_ulo
 PFNGLXGETSELECTEDEVENTSGIXPROC = CFUNCTYPE(None, POINTER(Display), GLXDrawable, POINTER(c_ulong)) 	# GL/glxext.h:486
 # SGI_cushion (GL/glxext.h:489)
 GLX_SGI_cushion = 1 	# GL/glxext.h:490
-Window = pyglet.libs.x11.xlib.Window
+Window = pyglet.libs.linux.x11.xlib.Window
 # GL/glxext.h:492
 glXCushionSGI = _link_function('glXCushionSGI', None, [POINTER(Display), Window, c_float], 'SGI_cushion')
 
@@ -422,7 +422,7 @@ glXCopySubBufferMESA = _link_function('glXCopySubBufferMESA', None, [POINTER(Dis
 PFNGLXCOPYSUBBUFFERMESAPROC = CFUNCTYPE(None, POINTER(Display), GLXDrawable, c_int, c_int, c_int, c_int) 	# GL/glxext.h:554
 # MESA_pixmap_colormap (GL/glxext.h:557)
 GLX_MESA_pixmap_colormap = 1 	# GL/glxext.h:558
-Colormap = pyglet.libs.x11.xlib.Colormap
+Colormap = pyglet.libs.linux.x11.xlib.Colormap
 # GL/glxext.h:560
 glXCreateGLXPixmapMESA = _link_function('glXCreateGLXPixmapMESA', GLXPixmap, [POINTER(Display), POINTER(XVisualInfo), Pixmap, Colormap], 'MESA_pixmap_colormap')
