+++ pyglet/input/linux/x11_xinput.py
@@ -2,11 +2,11 @@ import ctypes
 import pyglet
 from pyglet.input.base import Device, DeviceOpenException
 from pyglet.input.base import Button, RelativeAxis, AbsoluteAxis
-from pyglet.libs.x11 import xlib
+from pyglet.libs.linux.x11 import xlib
 from pyglet.util import asstr
 
 try:
-    from pyglet.libs.x11 import xinput as xi
+    from pyglet.libs.linux.x11 import xinput as xi
 
     _have_xinput = True
 except:
