+++ pyglet/libs/linux/x11/cursorfont.py
