+++ pyglet/libs/linux/x11/xinerama.py
@@ -30,7 +30,7 @@ class c_void(Structure):
     _fields_ = [('dummy', c_int)]
 
 
-import pyglet.libs.x11.xlib
+import pyglet.libs.linux.x11.xlib
 
 class struct_anon_93(Structure):
     __slots__ = [
@@ -49,7 +49,7 @@ struct_anon_93._fields_ = [
 ]
 
 XineramaScreenInfo = struct_anon_93 	# /usr/include/X11/extensions/Xinerama.h:40
-Display = pyglet.libs.x11.xlib.Display
+Display = pyglet.libs.linux.x11.xlib.Display
 # /usr/include/X11/extensions/Xinerama.h:44
 XineramaQueryExtension = _lib.XineramaQueryExtension
 XineramaQueryExtension.restype = c_int
