+++ pyglet/libs/linux/x11/xinput.py
@@ -30,7 +30,7 @@ class c_void(Structure):
     _fields_ = [('dummy', c_int)]
 
 
-import pyglet.libs.x11.xlib
+import pyglet.libs.linux.x11.xlib
 
 sz_xGetExtensionVersionReq = 8 	# /usr/include/X11/extensions/XI.h:56
 sz_xGetExtensionVersionReply = 32 	# /usr/include/X11/extensions/XI.h:57
@@ -241,10 +241,10 @@ class struct_anon_94(Structure):
         'first_axis',
         'axis_data',
     ]
-Display = pyglet.libs.x11.xlib.Display
-Window = pyglet.libs.x11.xlib.Window
-XID = pyglet.libs.x11.xlib.XID
-Time = pyglet.libs.x11.xlib.Time
+Display = pyglet.libs.linux.x11.xlib.Display
+Window = pyglet.libs.linux.x11.xlib.Window
+XID = pyglet.libs.linux.x11.xlib.XID
+Time = pyglet.libs.linux.x11.xlib.Time
 struct_anon_94._fields_ = [
     ('type', c_int),
     ('serial', c_ulong),
@@ -686,7 +686,7 @@ class struct_anon_111(Structure):
         'num_syms_supported',
         'syms_supported',
     ]
-KeySym = pyglet.libs.x11.xlib.KeySym
+KeySym = pyglet.libs.linux.x11.xlib.KeySym
 struct_anon_111._fields_ = [
     ('class', XID),
     ('length', c_int),
@@ -1113,7 +1113,7 @@ class struct__XDeviceInfo(Structure):
         'use',
         'inputclassinfo',
     ]
-Atom = pyglet.libs.x11.xlib.Atom
+Atom = pyglet.libs.linux.x11.xlib.Atom
 struct__XDeviceInfo._fields_ = [
     ('id', XID),
     ('type', Atom),
@@ -1427,7 +1427,7 @@ XDeviceBell = _lib.XDeviceBell
 XDeviceBell.restype = c_int
 XDeviceBell.argtypes = [POINTER(Display), POINTER(XDevice), XID, XID, c_int]
 
-KeyCode = pyglet.libs.x11.xlib.KeyCode
+KeyCode = pyglet.libs.linux.x11.xlib.KeyCode
 # /usr/include/X11/extensions/XInput.h:5878
 XGetDeviceKeyMapping = _lib.XGetDeviceKeyMapping
 XGetDeviceKeyMapping.restype = POINTER(KeySym)
@@ -1438,7 +1438,7 @@ XChangeDeviceKeyMapping = _lib.XChangeDeviceKeyMapping
 XChangeDeviceKeyMapping.restype = c_int
 XChangeDeviceKeyMapping.argtypes = [POINTER(Display), POINTER(XDevice), c_int, c_int, POINTER(KeySym), c_int]
 
-XModifierKeymap = pyglet.libs.x11.xlib.XModifierKeymap
+XModifierKeymap = pyglet.libs.linux.x11.xlib.XModifierKeymap
 # /usr/include/X11/extensions/XInput.h:5899
 XGetDeviceModifierMapping = _lib.XGetDeviceModifierMapping
 XGetDeviceModifierMapping.restype = POINTER(XModifierKeymap)
@@ -1544,7 +1544,7 @@ XGetDeviceDontPropagateList = _lib.XGetDeviceDontPropagateList
 XGetDeviceDontPropagateList.restype = POINTER(XEventClass)
 XGetDeviceDontPropagateList.argtypes = [POINTER(Display), Window, POINTER(c_int)]
 
-XEvent = pyglet.libs.x11.xlib.XEvent
+XEvent = pyglet.libs.linux.x11.xlib.XEvent
 # /usr/include/X11/extensions/XInput.h:6014
 XSendExtensionEvent = _lib.XSendExtensionEvent
 XSendExtensionEvent.restype = c_int
