+++ pyglet/libs/linux/x11/xlib.py
