+++ pyglet/libs/linux/x11/xrender.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import warnings
 from ctypes import Structure, c_ushort, c_ulong, c_int, c_void_p, POINTER
 import pyglet.lib
-from pyglet.libs.x11.xlib import Visual
+from pyglet.libs.linux.x11.xlib import Visual
 
 try:
     lib = pyglet.lib.load_library("Xrender")
