+++ pyglet/libs/linux/x11/xsync.py
